backups/containers/backup_20251111_204029/bot_app/requirements.txt
python-telegram-bot>=21
aiohttp>=3.9
httpx>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0
asyncpg